refactor: separate packages for xkcd and web handlers
