+++ cmd/web/main.go
@@ -1,14 +1,15 @@
 package main
 
 import (
+	"courses/server/handler/web"
 	"log"
 	"net/http"
 )
 
 func main() {
 	mux := http.NewServeMux()
-	mux.HandleFunc("/login", loginHandler)
-	mux.HandleFunc("/comics", comicsFinder)
+	mux.HandleFunc("/login", web.LoginHandler)
+	mux.HandleFunc("/comics", web.ComicsFinder)
 	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		http.Redirect(w, r, "/login", http.StatusMovedPermanently)
 	})
