+++ server/handler/xkcd/handlers_test.go
@@ -1,4 +1,4 @@
-package handler
+package xkcd
 
 import (
 	"bytes"
