+++ server/handler/xkcd/token_test.go
@@ -1,4 +1,4 @@
-package handler
+package xkcd
 
 import (
 	"testing"
